# Speed up linting and formatting by using more specific paths